+++ CHANGELOG.md
@@ -7,6 +7,8 @@ v0.3.0 (in development)
 - CLI: Do not emit a redundant error message when `git remote get-url` fails
 - Drop serde, serde-json, fancy-regex, and lazy-static dependencies
 - Remove the `GH_OWNER_RGX` and `GH_NAME_RGX` constants
+- Properly follow RFC 3986 when parsing username & password fields in
+  `www.github.com` URLs
 
 v0.2.1 (2022-10-19)
 -------------------